refactor: consolidate types to single file
